IOS-3592 Add container
