+++ Anytype/Sources/PresentationLayer/Flows/HomeChatCoordinator/HomeChatCoordinatorView.swift
@@ -1,0 +1,17 @@
+import SwiftUI
+import AnytypeCore
+import Services
+
+struct HomeChatCoordinatorView: View {
+    
+    @StateObject private var model: HomeWidgetsCoordinatorViewModel
+    @Environment(\.pageNavigation) private var pageNavigation
+    
+    init(spaceInfo: AccountInfo) {
+        self._model = StateObject(wrappedValue: HomeWidgetsCoordinatorViewModel(spaceInfo: spaceInfo))
+    }
+    
+    var body: some View {
+        Text("Chat Coordinator")
+    }
+}
